// src/pages/types.ts
export interface IMonsters {
    id: number;
    title: string;
    price: number;
    image: string;
    count: number
  }

  export interface ITours {
    id: number;
    title: string;
    price: number;
    image: string;
    count: number
  }

  export interface IWitchers {
    id: number
    title: string;
    image: string;
    price: number;
    count: number;
  }